requirements.txt
Flask>=3.0.0
flask-socketio>=5.3.0
python-whois>=0.9.0
eventlet>=0.33.0
python-telegram-bot>=22.0.0
gunicorn>=21.0.0